% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRCD.R
\name{mrcd}
\alias{mrcd}
\title{mrcd ...}
\usage{
mrcd(
  mX,
  target = 0,
  h = NULL,
  alpha = 0.75,
  rho = NULL,
  rescale = TRUE,
  rescaleSVD = TRUE,
  bc = FALSE,
  maxcond = 1000,
  minscale = 0.001,
  mindet = 0,
  objective = "geom",
  maxit = 200,
  initrho = 0.1
)
}
\arguments{
\item{mX}{...}

\item{target}{...}

\item{h}{...}

\item{alpha}{...}

\item{rho}{...}

\item{rescale}{...}

\item{rescaleSVD}{...}

\item{bc}{...}

\item{maxcond}{...}

\item{minscale}{...}

\item{mindet}{...}

\item{objective}{...}

\item{maxit}{...}

\item{initrho}{...}
}
\description{
mrcd ...
}
