% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRegCovD-package.r
\docType{package}
\name{MRegCovD-package}
\alias{MRegCovD}
\alias{MRegCovD-package}
\title{MRegCovD: TODO:title}
\description{
To learn more about simlab, start with the vignettes:
\code{browseVignettes(package = "MRegCovD")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.unimib.it}
  \item \url{https://github.com/giannamonti/MRegCovD}
  \item Report bugs at \url{https://github.com/giannamonti/MRegCovD/issues}
}

}
\author{
\strong{Maintainer}: Gianna Monti \email{gianna.monti@unimib.it}

}
\keyword{internal}
